% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichierR.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles distinctes}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Data frame contenant les données de trajets avec une colonne "Numéro de boucle"}
}
\value{
Un nombre (integer) représentant le nombre de boucles distinctes
}
\description{
Compter le nombre de boucles distinctes
}
\examples{
# compter_nombre_boucle(mes_donnees)
}
