% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichierR.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Data frame contenant les données de trajets avec une colonne Total}
}
\value{
Un nombre (integer) représentant la somme totale des trajets
}
\description{
Compter le nombre total de trajets
}
\examples{
# compter_nombre_trajets(mes_donnees)
}
