% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Données de comptage vélo Nantes Métropole}
\format{
Un data frame
}
\source{
\url{https://data.nantesmetropole.fr/}
}
\usage{
df_velo
}
\description{
Comptages des passages de vélos sur les boucles de Nantes Métropole.
}
\keyword{datasets}
