% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichierR.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtre les trajets sans anomalies}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Data frame contenant les données de trajets vélo}
}
\value{
Data frame filtré sans les lignes avec anomalies
}
\description{
Filtre les trajets sans anomalies
}
\examples{
\dontrun{
data_propre <- filtre_anomalie(data_velo)
}
}
