% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_trajet.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(df_velo, boucle = NULL)
}
\arguments{
\item{boucle}{Vecteur de numéros de boucle à sélectionner}

\item{trajet}{Data frame contenant les données de trajets}
}
\value{
Data frame filtré avec les boucles sélectionnées
}
\description{
Filtrer les trajets par numéro de boucle
}
\examples{
# filtrer_trajet(df_velo, c("880", "881"))
}
