% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichierR.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Graphique de la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Data frame contenant les données de trajets}
}
\value{
Un objet ggplot représentant la distribution des trajets par jour de la semaine
}
\description{
Graphique de la distribution hebdomadaire des trajets
}
\examples{
# plot_distribution_semaine(mes_donnees)
}
