% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichierR.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le maximum de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Data frame contenant les données de trajets avec une colonne Total}
}
\value{
Un data frame avec les colonnes "Boucle de comptage", "Jour" et "Total" pour le trajet maximum
}
\description{
Trouver le trajet avec le maximum de passages
}
\examples{
# trouver_trajet_max(mes_donnees)
}
